# openapi.yml
openapi: 3.0.0
info:
  title: Polyflix video service API
  description: Video Servie OpenAPI document.
  version: '2.0'
  contact: { }
tags: []
servers:
  - url: /v2.0.0
    description: v2
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    Video:
      type: object
      properties:
        slug:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        likes:
          type: number
        views:
          type: number
      required:
        - slug
        - description
        - thumbnail
paths:
  /videos:
    get:
      operationId: videocontroller-findall
      summary: Get a list of videos
      description: Get a list of videos
      parameters:
        - name: page
          required: true
          in: query
          schema:
            type: number
        - name: pageSize
          required: true
          in: query
          schema:
            type: number
        - name: order
          required: true
          in: query
          schema:
            type: string
        - name: slug
          required: false
          in: query
          schema:
            type: string
        - name: title
          required: false
          in: query
          schema:
            type: string
        - name: authorId
          required: false
          in: query
          schema:
            type: string
          description: Not implemented yet
        - name: collectionId
          required: false
          in: query
          schema:
            type: string
          description: Not implemented yet
        - name: visibility
          required: true
          in: query
          schema:
            enum:
              - PUBLIC
              - PROTECTED
              - PRIVATE
            type: string
          description: Not implemented yet
        - name: draft
          required: false
          in: query
          schema:
            type: boolean
          description: Not implemented yet
        - name: isWatched
          required: false
          in: query
          schema:
            type: boolean
          description: Not implemented yet
        - name: watchCount
          required: false
          in: query
          schema:
            type: number
          description: Not implemented yet
        - name: isWatching
          required: false
          in: query
          schema:
            type: boolean
          description: Not implemented yet
        - name: joinWithPublisher
          required: false
          in: query
          schema:
            type: boolean
          description: Not implemented yet
        - name: exact
          required: false
          in: query
          schema:
            type: boolean
          description: Not implemented yet
        - name: minViews
          required: false
          in: query
          schema:
            type: number
        - name: maxViews
          required: false
          in: query
          schema:
            type: number
        - name: startDate
          required: false
          in: query
          schema:
            format: date-time
            type: string
          description: Not implemented yet
        - name: endDate
          required: false
          in: query
          schema:
            format: date-time
            type: string
          description: Not implemented yet
        - name: tags
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
          description: Not implemented yet
        - name: allTags
          required: false
          in: query
          schema:
            default: false
            type: boolean
          description: Not implemented yet
      responses:
        undefined:
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Video'
      tags:
        - Video
      security:
        - bearer: []
    parameters: []
  '/videos/{slug}':
    get:
      operationId: videocontroller-findone
      summary: Get a video by slug
      description: Get a video by slug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Resource not found
      tags:
        - Video
      security:
        - bearer: []
    parameters:
      - schema:
          type: string
        name: slug
        in: path
        required: true
  '/stats/video/{id}/like':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    patch:
      summary: ''
      operationId: patch-stats-like
      responses:
        '200':
          description: OK
      description: Like/Unlike video by slug
      security:
        - bearer: []
  '/api/videos/metadata/{id}':
    get:
      operationId: videocontroller-findmetas
      summary: Get a video metadata on youtube data API
      description: Get a video metadata on youtube data API
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Resource not found
      tags:
        - Video
  '/video/{slug}/token':
    get:
      operationId: videocontroller-videoPSU
      summary: Get a MinIO PSU token for video by slug
      description: Get a MinIO PSU token for video by slug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Resource not found
        '418':
          description: Cannot get a token from an external video
        '401':
          description: You cannot access this video or You cannot create a token for this video
      tags:
        - Video
      security:
        - bearer: []
    parameters:
      - schema:
          type: string
        name: slug
        in: path
        required: true